Cleanup, asyncio client and replicaset

+++ dockerdb/__init__.py
@@ -104,37 +104,72 @@ class Mongo(Service):
     name = 'mongo'
     mongo_port = 27017
 
-    def __init__(self, tag, wait=False, port=27017, **kwargs):
+    def __init__(self, tag, wait=False, port=27017, replicaset=None, **kwargs):
         self.port = port
 
+        if replicaset is True:
+            replicaset = 'rs0'
+
+        self.replicaset_ready = False
+        self.replicaset = replicaset
+
+        if replicaset:
+            kwargs['command'] = ['mongod', '--replSet', replicaset]
+
         ports = {'{}/tcp'.format(self.mongo_port): ('127.0.0.1', self.port)}
-        Service.__init__(self, 'mongo:' + tag, ports=ports)
+        Service.__init__(self, 'mongo:' + tag, ports=ports, **kwargs)
         if wait:
             self.wait()
 
     def check_ready(self):
         """Check if something responds to ``url``."""
-        import pymongo
         import pymongo.errors
 
         client = self.pymongo_client()
         try:
-            client.admin.command('ismaster')
-            return True
+            is_master = client.admin.command('ismaster')
         except pymongo.errors.ConnectionFailure:
             return False
 
-    def pymongo_client(self):
-        # lazy load pymongo
-        import pymongo
-        import pymongo.errors
+        if self.replicaset and not self.replicaset_ready:
+            host = '{}:{}'.format(self.ip_address(), self.port)
+            conf = {
+                '_id': self.replicaset,
+                'members': [{'_id': 0, 'host': host}]
+            }
+            try:
+                client.admin.command('replSetInitiate', conf)
+                self.replicaset_ready = True
+            except pymongo.errors.OperationFailure:
+                # already initlized
+                self.replicaset_ready = True
+            except pymongo.errors.NetworkTimeout:
+                # for some reason this likes to time out
+                return False
 
-        server = self.ip_address()
-        return pymongo.MongoClient(
-            server, self.port, socketTimeoutMS=100, connectTimeoutMS=100)
+        if self.replicaset:
+            return is_master.get('ismaster', False)
+        return True
+
+    def client_args(self):
+        host = '{}:{}'.format(self.ip_address(), self.port)
+        return {
+            'host': [host],
+            'socketTimeoutMS': 200,
+            'connectTimeoutMS': 200
+        }
+
+    def pymongo_client(self):
+        import pymongo
+        return pymongo.MongoClient(**self.client_args())
+
+    def asyncio_client(self):
+        import motor.motor_asyncio
+        return motor.motor_asyncio.AsyncIOMotorClient(**self.client_args())
 
     def factory_reset(self):
         """factory reset the database"""
         client = self.pymongo_client()
         for db in client.database_names():
-            client.drop_database(db)
+            if db not in ('admin', 'config', 'local'):
+                client.drop_database(db)
